+++ src/Extensions/OptionStringExtensions.cs
@@ -2,12 +2,7 @@ namespace Ametrin.Optional;
 
 public static class OptionStringExtensions
 {
-    // public static Option<T> ParseOrNone<T>(this string s, IFormatProvider? provider = null) where T : ISpanParsable<T>
-    //     => T.TryParse(s, provider, out var result) ? Option.Success(result) : default;
-    // public static Option<T> ParseOrNone<T>(this ReadOnlySpan<char> s, IFormatProvider? provider = null) where T : ISpanParsable<T>
-    //     => T.TryParse(s, provider, out var result) ? Option.Success(result) : default;
-
-    public static Option<T> Parse<T>(this Option<string> option, IFormatProvider? provider = null) where T : ISpanParsable<T>
+    public static Option<T> Parse<T>(this Option<string> option, IFormatProvider? provider = null) where T : IParsable<T>
         => option._hasValue && T.TryParse(option._value, provider, out var result) ? Option.Success(result) : default;
 
     public static Option<string> WhereNotEmpty(this Option<string> option)
@@ -15,5 +10,4 @@ public static class OptionStringExtensions
 
     public static Option<string> WhereNotWhiteSpace(this Option<string> option)
         => option.WhereNot(string.IsNullOrWhiteSpace);
-
 }