+++ changelog.md
@@ -1,2 +1,4 @@
 ## 0.0.2 (unreleased)
 - update TUnit to 0.18.60
+- Consume returns reduced Option
+- use custom Exceptions for `.OrThrow()`